dynamic_cache/requirements.txt
fastapi
uvicorn
google-generativeai
faiss-cpu
numpy
pydantic
python-dotenv
redis
aioredis
httpx
streamlit
plotly
pandas
requests
